Keep transcode methods concrete
