+++ conda_lockfiles/rattler_lock/v6.py
@@ -13,6 +13,7 @@ from conda.models.environment import (
     Environment,
     EnvironmentConfig,
 )
+from conda.plugins.types import EnvironmentSpecBase
 from pydantic import (
     BaseModel,
     Field,
@@ -26,7 +27,6 @@ from ruamel.yaml.parser import ParserError
 from .. import CONDA_PYPI_CHANNEL_NAME, PYTHONHOSTED_URL_PREFIX
 from ..exceptions import CondaLockfilesParserError, CondaLockfilesValidationError
 from ..load_yaml import load_yaml
-from ..lockfile import LockfileSpecBase
 from ..records_from_conda_urls import _records_for_export, records_from_conda_urls
 from ..validate_urls import validate_urls
 
@@ -362,7 +362,7 @@ def multiplatform_export(envs: Iterable[Environment]) -> str:
         ) from e
 
 
-class RattlerLockV6Loader(LockfileSpecBase):
+class RattlerLockV6Loader(EnvironmentSpecBase):
     detection_supported: ClassVar[bool] = True
 
     def __init__(self, path: PathType):
@@ -425,11 +425,22 @@ class RattlerLockV6Loader(LockfileSpecBase):
             )
         return rattler_lock_v6_to_conda_env(self._model, platform=platform)
 
-    def _environments_for_transcode(
+    def transcode(
         self,
-        platforms: tuple[str, ...],
-        target_format: str,
-    ) -> list[Environment]:
+        platforms: Iterable[str],
+        *,
+        format_name: str,
+    ) -> str:
+        """Render selected platforms without fetching package artifacts."""
+        requested = tuple(platforms)
+        if not requested:
+            raise CondaValueError("At least one platform is required for transcoding.")
+        missing = sorted(set(requested) - set(self.available_platforms))
+        if missing:
+            raise CondaValueError(
+                f"Platform(s) not in lockfile: {', '.join(missing)}. "
+                f"Available platforms: {', '.join(self.available_platforms)}"
+            )
         if set(self._model.environments) != {"default"}:
             raise CondaValueError(
                 "Cannot transcode a rattler-lock-v6 file with multiple environments "
@@ -437,7 +448,7 @@ class RattlerLockV6Loader(LockfileSpecBase):
             )
         references = [
             reference
-            for platform in platforms
+            for platform in requested
             for reference in self._model.environments["default"].packages[platform]
         ]
         if any(reference.pypi for reference in references):
@@ -445,23 +456,33 @@ class RattlerLockV6Loader(LockfileSpecBase):
                 "Cannot transcode rattler-lock-v6 PyPI packages without losing "
                 "lockfile data."
             )
-        if target_format != FORMAT:
+        if format_name in (FORMAT, *ALIASES):
+            export = multiplatform_export
+            to_conda_lock = False
+        else:
             from ..conda_lock import v1 as conda_lock_v1
 
-            if target_format == conda_lock_v1.FORMAT and any(
-                reference.conda and reference.url.startswith(PYTHONHOSTED_URL_PREFIX)
-                for reference in references
-            ):
+            if format_name not in (conda_lock_v1.FORMAT, *conda_lock_v1.ALIASES):
                 raise CondaValueError(
-                    "Cannot transcode conda-pypi wheel records from rattler-lock-v6 "
-                    "to conda-lock-v1 without package metadata."
+                    f"Unsupported lockfile transcode format: {format_name}"
                 )
-        return [
+            export = conda_lock_v1.multiplatform_export
+            to_conda_lock = True
+        if to_conda_lock and any(
+            reference.conda and reference.url.startswith(PYTHONHOSTED_URL_PREFIX)
+            for reference in references
+        ):
+            raise CondaValueError(
+                "Cannot transcode conda-pypi wheel records from rattler-lock-v6 "
+                "to conda-lock-v1 without package metadata."
+            )
+        envs = [
             _rattler_lock_v6_to_conda_env(
                 self._model,
                 "default",
                 platform,
                 fetch=False,
             )
-            for platform in platforms
+            for platform in requested
         ]
+        return export(envs)
